requirements.txt
absl-py==2.3.1
alembic==1.16.4
astunparse==1.6.3
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
flask==2.3.3
flask-login==0.6.3
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
mako==1.3.10
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
opencv-python==4.12.0.88
opt-einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
pip==25.0.1
protobuf==5.29.5
pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
sqlalchemy==2.0.41
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
typing-extensions==4.14.1
urllib3==2.5.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wtforms==3.2.1
